Texture converter, DevIL: ABGR DDS support. Textures: converted ARGB to ABGR. Misc: warning fixes, etc.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3631 3db68df2-c116-0410-a063-a993310a9797

+++ source/renderer/WaterManager.h
@@ -38,7 +38,7 @@ public:
 	float m_TWaterTrans;
 	float m_SWaterScrollCounter;
 	float m_TWaterScrollCounter;
-	float m_WaterTexTimer;
+	double m_WaterTexTimer;
 
 public:
 	WaterManager();
